requirements.txt
PyYAML
psycopg2
sqlalchemy
pandas
kaggle